+++ src/step1-index.js
@@ -1,4 +1,4 @@
-import { ERROR, INPUT, OUTPUT, RETRY_STRING } from "./constants/message.js";
+import { ERROR, INPUT, OUTPUT, RETRY_STRING } from "./view/message.js";
 import Input from "./view/Input.js";
 import Output from "./view/Output.js";
 import { divideByUnit } from "./utils/count.js";
@@ -69,10 +69,10 @@ const start = async () => {
 
   const retry = await Input.retry(async () => {
     const input = await Input.readLineAsync(INPUT.RETRY);
-    if (!RETRY_STRING.includes(input)) {
+    if (![RETRY_STRING.Yes, RETRY_STRING.No].includes(input.toUpperCase())) {
       throwError(ERROR.INVALID_RETRY_STRING);
     }
-    return input.toLowerCase();
+    return input.toUpperCase();
   });
 
   if (retry === "y") await game();
