+++ src/utils/throwError.js
@@ -1,4 +1,4 @@
-import { ERROR_PREFIX } from "../constants/message.js";
+import { ERROR_PREFIX } from "../view/message.js";
 
 export const throwError = (message) => {
   throw new Error(`${ERROR_PREFIX} ${message}`);
